:truck: api: move get-private-certificate usecase from lib to src
